Implement permanent action confirmation component

+++ src/features/images/DeleteImagesAlert.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React, { useState, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { red } from '@radix-ui/colors';
 import {
@@ -32,6 +32,7 @@ import {
 } from '../tasks/tasksSlice.js';
 import { SimpleSpinner, SpinnerOverlay } from '../../components/Spinner.jsx';
 import { DeleteImagesProgressBar } from './DeleteImagesProgressBar.jsx';
+import PermanentActionConfirmation from '../../components/PermanentActionConfirmation.jsx';
 
 const DeleteImagesAlert = () => {
   const dispatch = useDispatch();
@@ -52,7 +53,7 @@ const DeleteImagesAlert = () => {
     }
   }, [deleteImagesTaskLoading, dispatch]);
 
-  const handleConfirmDelete = () => {
+  const handleDeleteClick = () => {
     if (alertState.deleteImagesAlertByFilter) {
       // if deleting by filter, always delete using task handler
       dispatch(deleteImagesTask({ imageIds: [], filters: filters }));
@@ -70,6 +71,8 @@ const DeleteImagesAlert = () => {
     dispatch(setDeleteImagesAlertStatus({ openStatus: false }));
     dispatch(clearDeleteImagesTask());
   };
+
+  const [confirmedDelete, setConfirmedDelete] = useState(false);
 
   const deleteByIdLimitExceeded =
     !alertState.deleteImagesAlertByFilter && selectedImages.length > ASYNC_IMAGE_DELETE_BY_ID_LIMIT;
@@ -99,7 +102,7 @@ const DeleteImagesAlert = () => {
   );
   const selectionText = (
     <div>
-      <p>This will delete all currently selected images. This action can not be undone.</p>
+      <p>This will delete all currently selected images. This action cannot be undone.</p>
     </div>
   );
 
@@ -160,21 +163,27 @@ const DeleteImagesAlert = () => {
             </SpinnerOverlay>
           )}
           <AlertTitle>{title}</AlertTitle>
-          {text}
+          <div>
+            {text}
+            <PermanentActionConfirmation
+              text="permanently delete"
+              setConfirmed={setConfirmedDelete}
+            />
+          </div>
           <div style={{ display: 'flex', gap: 25, justifyContent: 'flex-end' }}>
             <Button size="small" css={{ border: 'none' }} onClick={handleCancelDelete}>
               Cancel
             </Button>
             <Button
               size="small"
-              disabled={deleteByIdLimitExceeded || byFilterLimitExceeded}
+              disabled={deleteByIdLimitExceeded || byFilterLimitExceeded || !confirmedDelete}
               css={{
                 backgroundColor: red.red4,
                 color: red.red11,
                 border: 'none',
                 '&:hover': { color: red.red11, backgroundColor: red.red5 },
               }}
-              onClick={handleConfirmDelete}
+              onClick={handleDeleteClick}
             >
               Yes, delete
             </Button>
